agent: KSync::CreateVhostIntf fails on FreeBSD
In contrary to what the method name states, it actually is supposed
to put an interface in the UP state. This has been misunderstood
and FreeBSD version of code attempted to create the interface which,
at the point of method invocation, existed; this caused the method
to fail on assert.