agent: Disable destruction of flow device on FreeBSD
The flow device created by vRouter module is removed and recreated
by userspace Agent with the ID obtained from vRouter module; this
does not work correctly on FreeBSD due to difference in device
node creation. It is enouth to disable the destruction and use
the device created by vrouter module instead.